user docs for the OTel bridge, deprecate OpenTracing bridge
