+++ docs/opentracing.asciidoc
@@ -5,7 +5,10 @@ NOTE: For the best reading experience,
 please view this documentation at https://www.elastic.co/guide/en/apm/agent/nodejs/current/opentracing.html[elastic.co]
 endif::[]
 
-== OpenTracing API
+== OpenTracing bridge
+
+NOTE: https://opentracing.io/[OpenTracing] is discontinued in favor of OpenTelemetry. This Elastic APM OpenTracing bridge is **deprecated**. Consider using the <<opentelemetry-bridge>> instead.
+
 
 The Elastic APM OpenTracing bridge allows creating Elastic APM transactions and spans,
 using the https://opentracing-javascript.surge.sh/[OpenTracing API].
